hibid_scraper/pyproject.toml
[project]
name = "hibid-scraper"
version = "0.1.0"
description = "HiBid auction scraper that stores raw JSON payloads in SQLite"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
postgres = [
    "psycopg2-binary>=2.9.9",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
